requirements.txt
streamlit
supabase
pillow
python-dotenv
pandas
openpyxl